requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
spotipy==2.24.0
